requirements.txt
aiogram==2.18
asyncpg==0.26.0
SQLAlchemy==1.3.23
psycopg2-binary==2.8.6
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==1.0.0
qrcode==7.3
pyzbar==0.1.8
Pillow==9.4.0
requests==2.31.0
aiohttp==3.8.4
